fix(server): 지렁이 몸통 좌표 데이터 타입을 클라이언트에 맞게 수정
* as-is: (u64, u64)
* to-be: (f32, f32)

+++ src/network/mod.rs
@@ -13,7 +13,7 @@ pub struct WormBody {
 
     // color: ???
 
-    positions: Vec<(u64, u64)>
+    positions: Vec<(f32, f32)>
 }
 
 impl WormBody {
